requirements.txt
torch
torchvision
pandas
open_clip_torch
pyarrow
duckdb
sqlparse